feat: require admin for meeting deletion
