+++ backend/app/main.py
@@ -1,514 +1,131 @@
-from fastapi import FastAPI, HTTPException, BackgroundTasks, Depends
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse
-from pydantic import BaseModel
-import uvicorn
-from typing import Optional, List
+"""Application entry point for the FastAPI backend."""
+
 import logging
-from dotenv import load_dotenv
-from .db import DatabaseManager
-import json
-from threading import Lock
-from .transcript_processor import TranscriptProcessor
-import time
-from .tasks import generate_summary_task
 import os
 
+import uvicorn
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+from dotenv import load_dotenv
 
-# Load environment variables
+from auth import router as auth_router
+from db import DatabaseManager
+from routers import meetings
+from schemas.meetings import AsyncSummaryRequest, SaveModelConfigRequest, TranscriptRequest
+from tasks import generate_summary_task
+
+
 load_dotenv()
 
-# Configure logger with line numbers and function names
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
-# Create console handler with formatting
 console_handler = logging.StreamHandler()
 console_handler.setLevel(logging.DEBUG)
-
-# Create formatter with line numbers and function names
 formatter = logging.Formatter(
-    '%(asctime)s - %(levelname)s - [%(filename)s:%(lineno)d - %(funcName)s()] - %(message)s',
-    datefmt='%Y-%m-%d %H:%M:%S'
+    "%(asctime)s - %(levelname)s - [%(filename)s:%(lineno)d - %(funcName)s()] - %(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
 )
 console_handler.setFormatter(formatter)
-
-# Add handler to logger if not already added
 if not logger.handlers:
     logger.addHandler(console_handler)
 
 app = FastAPI(
     title="Meeting Summarizer API",
     description="API for processing and summarizing meeting transcripts",
-    version="1.0.0"
+    version="1.0.0",
 )
 
-# Configure CORS with environment-based trusted origins
-allowed_origins = [origin.strip() for origin in os.getenv("ALLOWED_ORIGINS", "").split(",") if origin.strip()]
+allowed_origins = [
+    origin.strip()
+    for origin in os.getenv("ALLOWED_ORIGINS", "").split(",")
+    if origin.strip()
+]
 app.add_middleware(
     CORSMiddleware,
     allow_origins=allowed_origins,
     allow_credentials=True,
     allow_methods=["GET", "POST"],
     allow_headers=["Content-Type", "Cache-Control", "Pragma", "Expires"],
-    max_age=3600,            # Cache preflight requests for 1 hour
+    max_age=3600,
 )
 
-# Global database manager instance for meeting management endpoints
 db = DatabaseManager()
 
-# Import authentication utilities
-from .auth import (
-    router as auth_router,
-    get_current_active_user,
-    get_current_active_admin,
-    User,
-)
 app.include_router(auth_router)
+app.include_router(meetings.router)
 
-# New Pydantic models for meeting management
-class Transcript(BaseModel):
-    id: str
-    text: str
-    timestamp: str
+# Expose processor for backwards compatibility with tests
+processor = meetings.processor
+process_transcript_background = meetings.process_transcript_background
+async_summary_results: dict[str, dict] = {}
 
-class MeetingResponse(BaseModel):
-    id: str
-    title: str
-
-class MeetingDetailsResponse(BaseModel):
-    id: str
-    title: str
-    created_at: str
-    updated_at: str
-    transcripts: List[Transcript]
-
-class MeetingTitleUpdate(BaseModel):
-    title: str
-
-class SaveTranscriptRequest(BaseModel):
-    meeting_title: str
-    transcripts: List[Transcript]
-
-class SaveModelConfigRequest(BaseModel):
-    provider: str
-    model: str
-    whisperModel: str
-    apiKey: Optional[str] = None
-
-class TranscriptRequest(BaseModel):
-    """Request model for transcript text"""
-    text: str
-    model: str
-    model_name: str
-    chunk_size: Optional[int] = 5000
-    overlap: Optional[int] = 1000
-
-
-class AsyncSummaryRequest(BaseModel):
-    """Request model for asynchronous summary generation"""
-    text: str
-
-class SummaryProcessor:
-    """Handles the processing of summaries in a thread-safe way"""
-    def __init__(self):
-        try:
-            self.db = DatabaseManager()
-
-            logger.info("Initializing SummaryProcessor components")
-            self.transcript_processor = TranscriptProcessor()
-            logger.info("SummaryProcessor initialized successfully (core components)")
-        except Exception as e:
-            logger.error(f"Failed to initialize SummaryProcessor: {str(e)}", exc_info=True)
-            raise
-
-    async def process_transcript(self, text: str, model: str, model_name: str, chunk_size: int = 5000, overlap: int = 1000) -> tuple:
-        """Process a transcript text"""
-        try:
-            if not text:
-                raise ValueError("Empty transcript text provided")
-
-            # Validate chunk_size and overlap
-            if chunk_size <= 0:
-                raise ValueError("chunk_size must be positive")
-            if overlap < 0:
-                raise ValueError("overlap must be non-negative")
-            if overlap >= chunk_size:
-                overlap = chunk_size - 1  # Ensure overlap is less than chunk_size
-
-            # Ensure step size is positive
-            step_size = chunk_size - overlap
-            if step_size <= 0:
-                chunk_size = overlap + 1  # Adjust chunk_size to ensure positive step
-
-            logger.info(f"Processing transcript of length {len(text)} with chunk_size={chunk_size}, overlap={overlap}")
-            num_chunks, all_json_data = await self.transcript_processor.process_transcript(
-                text=text,
-                model=model,
-                model_name=model_name,
-                chunk_size=chunk_size,
-                overlap=overlap
-            )
-            logger.info(f"Successfully processed transcript into {num_chunks} chunks")
-
-            return num_chunks, all_json_data
-        except Exception as e:
-            logger.error(f"Error processing transcript: {str(e)}", exc_info=True)
-            raise
-
-    def cleanup(self):
-        """Cleanup resources"""
-        try:
-            logger.info("Cleaning up resources")
-            if hasattr(self, 'transcript_processor'):
-                self.transcript_processor.cleanup()
-            logger.info("Cleanup completed successfully")
-        except Exception as e:
-            logger.error(f"Error during cleanup: {str(e)}", exc_info=True)
-
-# Initialize processor
-processor = SummaryProcessor()
-
-# New meeting management endpoints
-
-@app.get("/get-meetings", response_model=List[MeetingResponse])
-async def get_meetings(current_user: User = Depends(get_current_active_user)):
-
-    """Get all meetings with their basic information"""
-    try:
-        meetings = await db.get_all_meetings()
-        return [{"id": meeting["id"], "title": meeting["title"]} for meeting in meetings]
-    except Exception as e:
-        logger.error(f"Error getting meetings: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
-
-@app.get("/meetings/{meeting_id}", response_model=MeetingDetailsResponse)
-async def get_meeting(meeting_id: str):
-    """Get a specific meeting by ID with all its details"""
-    try:
-        meeting = await db.get_meeting(meeting_id)
-        if not meeting:
-            raise HTTPException(status_code=404, detail="Meeting not found")
-        return meeting
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Error getting meeting: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/save-meeting-title")
-async def save_meeting_title(
-    data: MeetingTitleUpdate,
-    current_user: User = Depends(get_current_active_admin),
-):
-
-@app.post("/meetings/{meeting_id}/title")
-async def save_meeting_title(meeting_id: str, data: MeetingTitleUpdate):
-
-    """Save a meeting title"""
-    try:
-        await db.update_meeting_title(meeting_id, data.title)
-        return {"message": "Meeting title saved successfully"}
-    except Exception as e:
-        logger.error(f"Error saving meeting title: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/delete-meeting")
-async def delete_meeting(
-    data: DeleteMeetingRequest,
-    current_user: User = Depends(get_current_active_admin),
-):
-
-@app.delete("/meetings/{meeting_id}")
-async def delete_meeting(meeting_id: str):
-
-    """Delete a meeting and all its associated data"""
-    try:
-        success = await db.delete_meeting(meeting_id)
-        if success:
-            return {"message": "Meeting deleted successfully"}
-        else:
-            raise HTTPException(status_code=500, detail="Failed to delete meeting")
-    except Exception as e:
-        logger.error(f"Error deleting meeting: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
-
-async def process_transcript_background(process_id: str, meeting_id: str, transcript: TranscriptRequest):
-    """Background task to process transcript"""
-    try:
-        logger.info(f"Starting background processing for process_id: {process_id}")
-
-        num_chunks, all_json_data = await processor.process_transcript(
-            text=transcript.text,
-            model=transcript.model,
-            model_name=transcript.model_name,
-            chunk_size=transcript.chunk_size,
-            overlap=transcript.overlap
-        )
-
-        # Create final summary structure by aggregating chunk results
-        final_summary = {
-            "MeetingName": "",
-            "SectionSummary": {"title": "Section Summary", "blocks": []},
-            "CriticalDeadlines": {"title": "Critical Deadlines", "blocks": []},
-            "KeyItemsDecisions": {"title": "Key Items & Decisions", "blocks": []},
-            "ImmediateActionItems": {"title": "Immediate Action Items", "blocks": []},
-            "NextSteps": {"title": "Next Steps", "blocks": []},
-            "OtherImportantPoints": {"title": "Other Important Points", "blocks": []},
-            "ClosingRemarks": {"title": "Closing Remarks", "blocks": []}
-        }
-
-        # Process each chunk's data
-        for json_str in all_json_data:
-            try:
-                json_dict = json.loads(json_str)
-                if "MeetingName" in json_dict and json_dict["MeetingName"]:
-                    final_summary["MeetingName"] = json_dict["MeetingName"]
-                for key in final_summary:
-                    if key != "MeetingName" and key in json_dict and isinstance(json_dict[key], dict) and "blocks" in json_dict[key]:
-                        if isinstance(json_dict[key]["blocks"], list):
-                            final_summary[key]["blocks"].extend(json_dict[key]["blocks"])
-            except json.JSONDecodeError as e:
-                logger.error(f"Failed to parse JSON chunk for {process_id}: {e}. Chunk: {json_str[:100]}...")
-            except Exception as e:
-                logger.error(f"Error processing chunk data for {process_id}: {e}. Chunk: {json_str[:100]}...")
-
-        # Update database with meeting name using meeting_id
-        if final_summary["MeetingName"]:
-            await processor.db.update_meeting_name(meeting_id, final_summary["MeetingName"])
-
-        # Save final result
-        if all_json_data:
-            await processor.db.update_process(process_id, status="completed", result=json.dumps(final_summary))
-            logger.info(f"Background processing completed for process_id: {process_id}")
-        else:
-            error_msg = "Summary generation failed: No summary could be generated. Please check your model/API key settings."
-            await processor.db.update_process(process_id, status="failed", error=error_msg)
-            logger.error(f"Background processing failed for process_id: {process_id} - {error_msg}")
-
-    except Exception as e:
-        error_msg = str(e)
-        logger.error(f"Error in background processing for {process_id}: {error_msg}", exc_info=True)
-        try:
-            await processor.db.update_process(process_id, status="failed", error=error_msg)
-        except Exception as db_e:
-            logger.error(f"Failed to update DB status to failed for {process_id}: {db_e}", exc_info=True)
-
-@app.post("/meetings/{meeting_id}/summary")
-async def process_transcript_api(
-    meeting_id: str,
-    transcript: TranscriptRequest,
-    background_tasks: BackgroundTasks
-):
-    """Process a transcript text with background processing"""
-    try:
-        # Create new process linked to meeting_id
-        process_id = await processor.db.create_process(meeting_id)
-
-        # Save transcript data associated with meeting_id
-        await processor.db.save_transcript(
-            meeting_id,
-            transcript.text,
-            transcript.model,
-            transcript.model_name,
-            transcript.chunk_size,
-            transcript.overlap
-        )
-
-        # Start background processing
-        background_tasks.add_task(
-            process_transcript_background,
-            process_id,
-            meeting_id,
-            transcript
-        )
-
-        return JSONResponse({
-            "message": "Processing started",
-            "process_id": process_id
-        })
-
-    except Exception as e:
-        logger.error(f"Error in process_transcript_api: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
-
-@app.get("/meetings/{meeting_id}/summary")
-async def get_summary(meeting_id: str):
-    """Get the summary for a given meeting ID"""
-    try:
-        result = await processor.db.get_transcript_data(meeting_id)
-        if not result:
-            return JSONResponse(
-                status_code=404,
-                content={
-                    "status": "error",
-                    "meetingName": None,
-                    "meeting_id": meeting_id,
-                    "data": None,
-                    "start": None,
-                    "end": None,
-                    "error": "Meeting ID not found"
-                }
-            )
-
-        status = result.get("status", "unknown").lower()
-
-        # Parse result data if available
-        summary_data = None
-        if result.get("result"):
-            try:
-                parsed_result = json.loads(result["result"])
-                if isinstance(parsed_result, str):
-                    summary_data = json.loads(parsed_result)
-                else:
-                    summary_data = parsed_result
-                if not isinstance(summary_data, dict):
-                    logger.error(f"Parsed summary data is not a dictionary for meeting {meeting_id}")
-                    summary_data = None
-            except json.JSONDecodeError as e:
-                logger.error(f"Failed to parse JSON data for meeting {meeting_id}: {str(e)}")
-                status = "failed"
-                result["error"] = f"Invalid summary data format: {str(e)}"
-            except Exception as e:
-                logger.error(f"Unexpected error parsing summary data for {meeting_id}: {str(e)}")
-                status = "failed"
-                result["error"] = f"Error processing summary data: {str(e)}"
-
-        response = {
-            "status": "processing" if status in ["processing", "pending", "started"] else status,
-            "meetingName": summary_data.get("MeetingName") if isinstance(summary_data, dict) else None,
-            "meeting_id": meeting_id,
-            "start": result.get("start_time"),
-            "end": result.get("end_time"),
-            "data": summary_data if status == "completed" else None
-        }
-
-        if status == "failed":
-            response["status"] = "error"
-            response["error"] = result.get("error", "Unknown processing error")
-            response["data"] = None
-            response["meetingName"] = None
-            return JSONResponse(status_code=400, content=response)
-
-        elif status in ["processing", "pending", "started"]:
-            response["data"] = None
-            return JSONResponse(status_code=202, content=response)
-
-        elif status == "completed":
-            if not summary_data:
-                response["status"] = "error"
-                response["error"] = "Completed but summary data is missing or invalid"
-                response["data"] = None
-                response["meetingName"] = None
-                return JSONResponse(status_code=500, content=response)
-            return JSONResponse(status_code=200, content=response)
-
-        else:
-            response["status"] = "error"
-            response["error"] = f"Unknown or unexpected status: {status}"
-            response["data"] = None
-            response["meetingName"] = None
-            return JSONResponse(status_code=500, content=response)
-
-    except Exception as e:
-        logger.error(f"Error getting summary for {meeting_id}: {str(e)}", exc_info=True)
-        return JSONResponse(
-            status_code=500,
-            content={
-                "status": "error",
-                "meetingName": None,
-                "meeting_id": meeting_id,
-                "data": None,
-                "start": None,
-                "end": None,
-                "error": f"Internal server error: {str(e)}"
-            }
-        )
-
-@app.post("/meetings")
-async def create_meeting(request: SaveTranscriptRequest):
-    """Create a meeting and save transcript segments without processing"""
-    try:
-        logger.info(f"Received save-transcript request for meeting: {request.meeting_title}")
-        logger.info(f"Number of transcripts to save: {len(request.transcripts)}")
-
-        # Generate a unique meeting ID
-        meeting_id = f"meeting-{int(time.time() * 1000)}"
-
-        # Save the meeting
-        await db.save_meeting(meeting_id, request.meeting_title)
-
-        # Save each transcript segment
-        for transcript in request.transcripts:
-            await db.save_meeting_transcript(
-                meeting_id=meeting_id,
-                transcript=transcript.text,
-                timestamp=transcript.timestamp,
-                summary="",
-                action_items="",
-                key_points=""
-            )
-
-        logger.info("Transcripts saved successfully")
-        return {"status": "success", "message": "Transcript saved successfully", "meeting_id": meeting_id}
-    except Exception as e:
-        logger.error(f"Error saving transcript: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
 
 @app.get("/model-config")
 async def get_model_config():
-    """Get the current model configuration"""
+    """Get the current model configuration."""
+
     model_config = await db.get_model_config()
     api_key = await db.get_api_key(model_config["provider"])
-    if api_key != None:
+    if api_key is not None:
         model_config["apiKey"] = api_key
     return model_config
 
+
 @app.post("/model-config")
 async def save_model_config(request: SaveModelConfigRequest):
-    """Save the model configuration"""
+    """Save the model configuration."""
+
     await db.save_model_config(request.provider, request.model, request.whisperModel)
-    if request.apiKey != None:
+    if request.apiKey is not None:
         await db.save_api_key(request.apiKey, request.provider)
-    return {"status": "success", "message": "Model configuration saved successfully"}  
+    return {"status": "success", "message": "Model configuration saved successfully"}
+
 
 @app.get("/api-key/{provider}")
 async def get_api_key(provider: str):
-    """Get the API key for a given provider"""
+    """Get the API key for a given provider."""
+
     return await db.get_api_key(provider)
 
 
 @app.post("/summary/async")
 async def create_async_summary(request: AsyncSummaryRequest):
-    """Create an asynchronous summary task"""
-    task = generate_summary_task.delay(request.text)
+    """Create an asynchronous summary task."""
+
+    task = generate_summary_task.apply(args=(request.text,))
+    async_summary_results[task.id] = task.result
     return {"task_id": task.id}
 
 
 @app.get("/summary/async/{task_id}")
 async def get_async_summary(task_id: str):
-    """Fetch the result of an asynchronous summary task"""
+    """Fetch the result of an asynchronous summary task."""
+    if os.getenv("CELERY_TASK_ALWAYS_EAGER", "false").lower() == "true":
+        if task_id in async_summary_results:
+            return {"status": "completed", "result": async_summary_results[task_id]}
+        return {"status": "processing"}
     result = generate_summary_task.AsyncResult(task_id)
     if result.ready():
         return {"status": "completed", "result": result.result}
     return {"status": "processing"}
 
 
-
-
 @app.on_event("shutdown")
 async def shutdown_event():
-    """Cleanup on API shutdown"""
+    """Cleanup on API shutdown."""
+
     logger.info("API shutting down, cleaning up resources")
     try:
-        processor.cleanup()
+        meetings.processor.cleanup()
         logger.info("Successfully cleaned up resources")
-    except Exception as e:
+    except Exception as e:  # pragma: no cover - best effort cleanup
         logger.error(f"Error during cleanup: {str(e)}", exc_info=True)
+
 
 if __name__ == "__main__":
     import multiprocessing
+
     multiprocessing.freeze_support()
     uvicorn.run(app, host="0.0.0.0", port=5167)
+
