+++ backend/app/schemas/__init__.py
@@ -1,0 +1,2 @@
+# Package marker for schema modules.
+
